Implement character limit for number input
Enhanced the handleNumber function to limit the input to a maximum of 6 digits. Now, when a number button is clicked, the script checks if the current input length is less than or equal to 5. If true, it adds the number to the currentValue and updates the currentScreen; otherwise, it displays an alert notifying the user that more than 6 digits are not allowed.